refactor(tests): move coverage gap tests into relevant test files
Distribute tests from test_coverage_gaps.py into their topical homes:
- test_weight_dimensions: TestRejectWeights, TestResolveVintageWeightToContext
- test_point: TestMapPerVintageEdgeCases
- test_parameter_validation: TestAggregationMethodNonStringList
- test_stepwise_vintagewise: TestCollapseVintageDimensionWithMeta
- test_interval: TestCollapseCoverageRatesEdgeCases, TestIntervalMultiVintageCoverage
- test_classification: TestRankingScorerCoverage
- test_scorer_checks: TestYieldScorerChecksIntegration

Remove test_coverage_gaps.py and all separator comments.
