+++ tests/metrics/test_parameter_validation.py
@@ -629,3 +629,17 @@ class TestMakeScorerGetScorer:
         from yohou.metrics import _SCORER_REGISTRY
 
         assert len(_SCORER_REGISTRY) == 27
+
+
+class TestAggregationMethodNonStringList:
+    """Cover aggregation_method list containing non-string elements."""
+
+    def test_non_string_list_element_raises(self):
+        """Non-string element in aggregation_method list raises ValueError."""
+        scorer = MeanAbsoluteError(aggregation_method=["stepwise", 42])
+        y_true = pl.DataFrame({
+            "time": [datetime(2020, 1, 1)],
+            "value": [1.0],
+        })
+        with pytest.raises(ValueError, match="All elements in aggregation_method must be strings"):
+            scorer.fit(y_true)
